// PlaybookAgent/src/generated/models/Jw_artifactsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Jw_artifactsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Jw_artifactsstatecode = keyof typeof Jw_artifactsstatecode;
export const Jw_artifactsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Jw_artifactsstatuscode = keyof typeof Jw_artifactsstatuscode;

export interface Jw_artifactsBase {
  importsequencenumber?: string;
  jw_artifactid: string;
  "jw_caseid@odata.bind"?: string;
  jw_name?: string;
  "jw_parentartifactid@odata.bind"?: string;
  jw_payload?: string;
  jw_referencekey?: string;
  jw_type: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Jw_artifactsstatecode;
  statuscode?: Jw_artifactsstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Jw_artifacts extends Jw_artifactsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  jw_caseidname?: string;
  jw_parentartifactidname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  jw_caseid?: object;
  _jw_caseid_value?: string;
  jw_parentartifactid?: object;
  _jw_parentartifactid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}